+++ src/main/java/com/entrego/dtos/OrderResponse.java
@@ -6,8 +6,9 @@ import com.entrego.domain.Product;
 import com.entrego.enums.OrderStatus;
 import com.entrego.enums.PaymentMethod;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public record OrderResponse(String id, String clientName, String numberOrder, Double total, LocalDateTime deliveryTime, OrderStatus status, Address address, PaymentMethod paymentMethod, List<ItemOrder> items) {
+public record OrderResponse(String id, String clientName, String numberOrder, BigDecimal total, LocalDateTime deliveryTime, OrderStatus status, Address address, PaymentMethod paymentMethod, List<ItemOrder> items) {
 }
